# openapi.yaml
openapi: 3.0.0
info:
  title: Messages API
  version: 1.0.0
#  you need add servers if you want to change default "localhost" ulr, in my case this is url because I use WLS for Windows11
servers:
  - url: http://172.17.23.239:8000/
paths:
  /messages:
    get:
      summary: Get Messages
      description: Retrieves a list of messages.
      parameters:
        - name: status
          in: query
          description: Filter messages by status
          required: false
          schema:
            type: "string"
            enum:
              - "sent"
              - "read"
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: "object"
                properties:
                  messages:
                    type: "array"
                    items:
                      type: "object"
                      properties:
                        uuid:
                          type: "string"
                        text:
                          type: "string"
                        status:
                          type: "string"
                          enum:
                            - "sent"
                            - "read"
              example:
                messages:
                  - text: "Hello, World!"
                    status: "read"
                  - text: "How are you today?"
                    status: "sent"
                  - text: "Thanks, I'm fine. How about you?"
                    status: "sent"
  /messages/send:
    get:
      summary: Send a Message
      parameters:
        - name: text
          in: query
          description: The text you want to send
          required: true
          schema:
            type: "string"
      responses:
        # - Please change status code to 200
        '204':
          description: Successfully sent

#  My code version with POST method
#  /messages/send:
#    post: # Change method to 'post'
#      summary: Send a Message
#      requestBody:
#        required: true
#        content:
#          application/json:
#            schema:
#              type: object
#              properties:
#                text:
#                  type: string
#      responses:
#        '200':
#          description: Successfully sent